tests/requirements.txt
pytest==7.4.0
pytest-asyncio==0.21.0
requests==2.31.0
kafka-python==2.0.2
docker==6.1.0
testcontainers==3.7.0